requirements.txt
beautifulsoup4
requests
markdownify
html2text
pandas
markdown
fastapi
scipy
uvicorn
transformers
torch
sqlalchemy
selenium
psycopg2-binary
python-dotenv